% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eq_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Takes raw NOAA data frame and returns a clean data frame}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{The uncleaned data.frame}
}
\value{
A clean data.frame
}
\description{
Takes raw NOAA data frame and returns a clean data frame
}
\details{
Date objects can only store positive years. BC years are marked with the IS_BC column
}
\examples{
\dontrun{
  raw_data <- readr::read_delim(raw_data_fn, "\\t"))
  clean_data <- eq_clean_data(raw_data)
}
}
