% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create an HTML label for the leaflet map}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{The input data.frame}
}
\value{
A vector of label strings
}
\description{
Create an HTML label for the leaflet map
}
\details{
This function should put together a character string for each earthquake that will show
    the cleaned location (as cleaned by the eq_location_clean() function created in Module 1),
    the magnitude (EQ_PRIMARY), and the total number of deaths (TOTAL_DEATHS), with boldface
    labels for each ("Location", "Total deaths", and "Magnitude"). If an earthquake is missing
    values for any of these, both the label and the value should be skipped for that element of
    the tag.
}
\examples{
\dontrun{
  data \%>\%
    dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
    eq_map(annot_col = "popup_text")
}
}
