% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eq_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean up the LOCATION_NAME column}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{The uncleaned data.frame}
}
\value{
A clean data.frame
}
\description{
Clean up the LOCATION_NAME column
}
\examples{
\dontrun{
  data <- uncleaned_data \%>\% eq_location_clean()
}
}
