% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Visualize earthquakes on a map}
\usage{
eq_map(data, annot_col = "DATE")
}
\arguments{
\item{data}{A data.frame containing the earthquake data}

\item{annot_col}{A column name indicating the column to be used for annotation pop up}
}
\description{
Visualize earthquakes on a map
}
\examples{
\dontrun{
  read_eq_clean_data() \%>\% eq_map()
}
}
