% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Custom Geom proto for the timeline}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\description{
Custom Geom proto for the timeline
}
\details{
This proto is for internal use only and not exported
}
\keyword{datasets}
