% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Specific them for our timelines}
\usage{
theme_timeline()
}
\value{
The theme for timeline goems
}
\description{
Specific them for our timelines
}
\examples{
\dontrun{
  data \%>\%
    ggplot2::gplot(ggplot2::aes(x=DATE, color = TOTAL_DEATHS, size = EQ_PRIMARY)) +
    geom_timeline() +
    theme_timeline()
}
}
