% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Custom Geom proto for timeline labesl}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 4.}
\usage{
GeomTimelineLabel
}
\description{
Custom Geom proto for timeline labesl
}
\details{
This proto is not exported
}
\keyword{datasets}
